+++ gzap/apworld/uzdoom/model/DoomWad.py
@@ -126,7 +126,10 @@ class DoomWad:
         """
         locations = []
         for map in world.maps:
-            locations += map.all_locations(world.spawn_filter, {})
+            if world.options.pretuning_mode:
+                locations += map.all_locations(None, {})
+            else:
+                locations += map.all_locations(world.spawn_filter, {})
         pool = DoomPool(self, locations, world)
         return pool